code/requirements.txt
functions-framework==3.6.0
google-cloud-storage==2.17.0
PyYAML==6.0.2
PyPDF2==3.0.1
reportlab==4.0.7
cloudevents==1.12.0
google-genai
